pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "factorlab"
version = "0.1.0"
description = "Global fundamental equity research platform"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.31",
    "pyyaml>=6.0",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "pandas>=2.1",
    "pyarrow>=14.0",
    "python-dotenv>=1.0",
    "sqlalchemy>=2.0",
    "psycopg[binary]>=3.1",
    "alembic>=1.13",
    "exchange-calendars>=4.5",
    "apscheduler>=3.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.1",
]
notebooks = [
    "jupyter>=1.0",
    "matplotlib>=3.8",
]
research = [
    "duckdb>=1.0",
    "polars>=1.0",
]
server = [
    "flask>=3.0",
    "gunicorn>=22.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/factorlab"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
src = ["src"]
line-length = 100
